requirements.txt
absl-py==2.3.1
astunparse==1.6.3
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
flatbuffers==25.9.23
fonttools==4.61.1
gast==0.7.0
google-pasta==0.2.0
grpcio==1.76.0
h5py==3.15.1
idna==3.11
joblib==1.5.2
keras==3.12.0
kiwisolver==1.4.9
libclang==18.1.1
logparser3==1.0.4
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
numpy==2.3.5
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
protobuf==6.33.2
Pygments==2.19.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2025.11.3
requests==2.32.5
rich==14.2.0
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
six==1.17.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorflow_cpu==2.20.0
termcolor==3.2.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.1
Werkzeug==3.1.4
wrapt==2.0.1
